Ease rpm requirement to allow smooth installation
This commit is meant to be reverted once RHEL 7.5 is released.

In the meantime, it allows at least package installation so testing
is possible.

The package effectively requires cockpit-ws >= 140 but RHEL 7.4 comes
with 138 only. Fine for Fedora >= 26.

This commit eases requirement in the .spec file to 138 but effectively
disables the service in start.sh when cockpit-ws <=140 is detected.