Exclude cockpit.conf from verification
In one of the preceding commits, cockpit.conf has been marked as a
config file.  However it is still reported as a changed file on
verification, although with the config file mark.  It confuses some
tools and since the file is supposed to be modified automatically, the
warning is useless anyway.  So let's exclude the file from
verification completely.

See https://bugzilla.redhat.com/1567936.